+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 * Added `::update_space!` command - [#2](https://github.com/contentful/contentful-database-importer.rb/issues/2)
 * Added `:skip_content_types` configuration option for `::update_space!`
+* Added `::query` support on `Resource` for selecting specific content for a table
 
 ## 0.1.0